requirements.txt
# AI Doctor — классификация родинок / меланома
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
Pillow==10.2.0
numpy==1.26.4
scipy>=1.11.0
tensorflow==2.16.1
kaggle==1.6.12
